requirements-dev.txt
Django~=3.1.4
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
Pillow~=8.1.0
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.3
